requirements.txt
Django==2.2.2
pycryptodome==3.8.2
pytz==2019.1
sqlparse==0.3.0
